requirements.txt
bitstring==3.1.9
ifaddr==0.1.7
lifxlan==1.2.7
Pillow==8.4.0
